refactor(deps): make the dependency index a CompilerHost capability
Address the altitude finding: bare-name dependency resolution was wired
only into `wado run`/`compile` (via CompilerOptions.dependencies built in
try_compile), so `wado check` and the LSP saw every bare-name import as
unresolved, and the path-relativization logic was duplicated in the CLI.

- Add `CompilerHost::dependency_index()` (default empty); the module
  loader seeds it once in `ModuleLoader::new`, so every entry point
  (compile, check, test, lsp) resolves `[dependencies]` uniformly.
- Implement it on `wado_lsp::FilesystemCompilerHost` (the shared host):
  nearest-manifest walk + path-dep entry resolution, relative to the
  host base. The CLI host delegates to it.
- Drop `CompilerOptions.dependencies`, `ModuleLoader::with_dependencies`,
  and the duplicated builder/path helpers in wado-cli.

Fixes `wado check` on bare-name dependencies (regression test added) and
the LSP false-unresolved-import; single resolution authority now.
