+++ wado-lsp/Cargo.toml
@@ -11,6 +11,7 @@ path = "src/bin/wado-lsp.rs"
 
 [dependencies]
 wado-compiler = { path = "../wado-compiler" }
+wado-manifest = { path = "../wado-manifest" }
 indexmap = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
